add triangle_mesh

+++ src/parallel_tiles_renderer.cpp
@@ -92,7 +92,7 @@ namespace raytracer11
 					}
 #ifdef WRITE_PER_THREAD_PERF_DATA
 					auto end_time = chrono::system_clock::now();
-					long tm = abs(chrono::duration_cast<chrono::nanoseconds>(end_time - start_time).count());
+					long long tm = abs(chrono::duration_cast<chrono::nanoseconds>(end_time - start_time).count());
 					if (tm < 0)cout << "X";
 					auto tmpt = (double)tm / (double)tiles_rendered;
 					cout << "thread " << this_thread::get_id() << " rendered " << tiles_rendered << " tiles"
@@ -123,7 +123,7 @@ namespace raytracer11
 					}
 #ifdef WRITE_PER_THREAD_PERF_DATA
 					auto end_time = chrono::system_clock::now();
-					long tm = abs(chrono::duration_cast<chrono::nanoseconds>(end_time - start_time).count());
+					long long tm = abs(chrono::duration_cast<chrono::nanoseconds>(end_time - start_time).count());
 					auto tmpt = (double)tm / (double)tiles_rendered;
 					cout << "thread " << this_thread::get_id() << " rendered " << tiles_rendered << " tiles"
 						<< " took " << (double)tm / 1000000.0 << "ms, " << tmpt / 1000000.0 << "ms/tile" << endl;
